refactor: ♻️ move shared constants module
