+++ src/utils/frontmatter-utils.ts
@@ -2,10 +2,10 @@
  * FrontmatterUtils.ts
  */
 
-import { FRONTMATTER_TO_ESCAPE } from 'constants/index';
 import { parseYaml } from 'obsidian';
 import { Frontmatter } from 'services/frontmatter';
 import type { ReadwiseEnvironment } from 'services/readwise-environment';
+import { FRONTMATTER_TO_ESCAPE } from 'src/constants';
 import { sampleMetadata } from 'test/sample-data';
 import { escapeMetadata } from 'utils/metadata-escape-utils';
