+++ tests/frontmatter-manager.spec.ts
@@ -3,7 +3,7 @@
  * Uses Jest with mocked obsidian module
  */
 
-import { DEFAULT_SETTINGS } from 'constants/index';
+import { DEFAULT_SETTINGS } from 'src/constants';
 import * as nunjucks from 'nunjucks';
 import { FileManager } from 'obsidian';
 import { FrontmatterManager } from 'services/frontmatter-manager';
